Move prototype plotter into plugin
