+++ OTAnalytics/plugin_prototypes/track_visualization/track_viz.py
@@ -1,3 +1,4 @@
+from abc import ABC, abstractmethod
 from typing import Iterable, Optional
 
 import numpy
@@ -10,7 +11,8 @@ from mpl_toolkits.axes_grid1 import Divider, Size
 from pandas import DataFrame
 from PIL import Image
 
-from OTAnalytics.application.plotting import TrackPlotter
+from OTAnalytics.application.datastore import Datastore
+from OTAnalytics.application.state import Plotter, TrackViewState
 from OTAnalytics.domain import track
 from OTAnalytics.domain.geometry import RelativeOffsetCoordinate
 from OTAnalytics.domain.section import Section
@@ -25,6 +27,65 @@ CLASS_PERSON = "person"
 CLASS_TRUCK = "truck"
 CLASS_BICYCLE = "bicycle"
 CLASS_TRAIN = "train"
+
+
+class TrackPlotter(ABC):
+    """
+    Abstraction to plot the background image.
+    """
+
+    @abstractmethod
+    def plot(
+        self,
+        tracks: Iterable[Track],
+        sections: Iterable[Section],
+        width: int,
+        height: int,
+        filter_classes: Iterable[str] = (
+            "car",
+            "motorcycle",
+            "person",
+            "truck",
+            "bicycle",
+            "train",
+        ),
+        num_min_frames: int = 30,
+        start_time: str = "",
+        end_time: str = "2022-09-15 07:05:00",
+        start_end: bool = True,
+        plot_sections: bool = True,
+        alpha: float = 0.1,
+        offset: Optional[RelativeOffsetCoordinate] = RelativeOffsetCoordinate(0, 0),
+    ) -> TrackImage:
+        pass
+
+
+class PlotterPrototype(Plotter):
+    def __init__(
+        self,
+        datastore: Datastore,
+        track_view_state: TrackViewState,
+        track_plotter: TrackPlotter,
+    ) -> None:
+        self._datastore = datastore
+        self._track_view_state = track_view_state
+        self._track_plotter = track_plotter
+
+    def plot(self) -> Optional[TrackImage]:
+        if track := next(iter(self._datastore.get_all_tracks())):
+            if new_image := self._datastore.get_image_of_track(track.id):
+                if self._track_view_state.show_tracks.get():
+                    track_image = self._track_plotter.plot(
+                        self._datastore.get_all_tracks(),
+                        self._datastore.get_all_sections(),
+                        width=new_image.width(),
+                        height=new_image.height(),
+                        offset=self._track_view_state.track_offset.get(),
+                    )
+                    return new_image.add(track_image)
+                else:
+                    return new_image
+        return None
 
 
 class MatplotlibTrackPlotter(TrackPlotter):
